Reorganize tera-ui-compoennts
