+++ libs/tera/ui-components/land-area/badges/kind/BadgeTeraLandAreaKind.vue
