+++ libs/tera/ui-components/land-area/card/index.ts
