+++ libs/tera/ui-components/land-task/badges/state/i18n/fr-FR.ts
